+++ app/src/main/java/org/stepic/droid/base/FragmentActivityBase.java
@@ -14,7 +14,7 @@ import org.stepic.droid.R;
 import org.stepic.droid.core.IShell;
 import org.stepic.droid.preferences.UserPreferences;
 import org.stepic.droid.store.operations.DatabaseManager;
-import org.stepic.droid.util.CoursePropertyResolver;
+import org.stepic.droid.util.resolvers.CoursePropertyResolver;
 import org.stepic.droid.util.resolvers.IStepResolver;
 
 import javax.inject.Inject;
